Move mime renderer to vegafusion package Add top-level vf.enable_mime()/enable_widget() methods
