+++ python/vegafusion-jupyter/vegafusion_jupyter/__init__.py
@@ -12,7 +12,7 @@ from .widget import VegaFusionWidget
 from ._version import __version__
 
 
-def enable_widget(
+def enable(
         download_source_link=None,
         debounce_wait=30,
         debounce_max_wait=60,
@@ -38,23 +38,9 @@ def enable_widget(
     )
 
 
-def enable_mime():
-    """
-    Enable the VegaFusion data transformer and renderer so that all Charts
-    are displayed using VegaFusion.
-
-    This isn't necessary in order to use the VegaFusionWidget directly
-    """
-    # Import vegafusion.transformer so that vegafusion-inline transform
-    # will be registered
-    import vegafusion.transformer
-    alt.renderers.enable('vegafusion-mime',)
-    alt.data_transformers.enable('vegafusion-inline')
-
-
 def disable():
     """
-    Disable the VegaFusion data transformer and renderer so that Charts
+    Disable the VegaFusion data transformers and renderers so that Charts
     are not displayed using VegaFusion
 
     Equivalent to
@@ -67,8 +53,8 @@ def disable():
 
     This does not affect the behavior of VegaFusionWidget
     """
-    alt.renderers.enable('default')
-    alt.data_transformers.enable('default')
+    from vegafusion import disable
+    disable()
 
 
 def _jupyter_labextension_paths():
