+++ python/vegafusion-jupyter/vegafusion_jupyter/renderer.py
@@ -6,7 +6,6 @@
 # this program the details of the active license.
 
 import altair as alt
-import vegafusion as vf
 
 def vegafusion_renderer(spec, **widget_options):
     """
@@ -21,14 +20,4 @@ def vegafusion_renderer(spec, **widget_options):
     display(widget)
     return {'text/plain': ""}
 
-def vegafusion_mime_renderer(spec):
-    import vl_convert as vlc
-    vega_spec = vlc.vegalite_to_vega(spec)
-    inline_datasets = vf.transformer.get_inline_datasets_for_spec(vega_spec)
-    tx_vega_spec, warnings = vf.runtime.pre_transform_spec(
-        vega_spec, "UTC", inline_datasets=inline_datasets
-    )
-    return {"application/vnd.vega.v5+json": tx_vega_spec}
-
 alt.renderers.register('vegafusion-widget', vegafusion_renderer)
-alt.renderers.register('vegafusion-mime', vegafusion_mime_renderer)
